services/ml-inference/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
numpy==1.26.3
scikit-learn==1.4.0
pandas==2.2.0
psycopg2-binary==2.9.9
joblib==1.3.2
